+++ packages/asc-ui/src/index.ts
@@ -67,7 +67,7 @@ import {
   MenuStyles,
   MenuButton,
 } from './components/Menu'
-import ThemeProvider from './theme'
+import ThemeProvider, { CompactThemeProvider } from './theme'
 import IconButton from './components/deprecated/IconButton'
 import ButtonBar from './components/deprecated/ButtonBar'
 import Radio from './components/Radio'
@@ -188,6 +188,7 @@ export {
   TextField,
   Theme,
   ThemeProvider,
+  CompactThemeProvider,
   TopBar,
   EditorialPost,
   Article,