+++ CHANGELOG.md
@@ -14,8 +14,15 @@ Prefix the change with one of these keywords:
 - _Fixed_: for any bug fixes.
 - _Security_: in case of vulnerabilities.
 
-## [0.13.0] - 2019-xx-xx
+## UNRELEASED
 
+## [0.13.2] - 2019-09-xx
+
+- Added: export CompactThemeProvider
+
+## [0.13.1] - 2019-09-02
+
+- Added: Checkbox component
 - Deprecated: Renamed the `color` function for extracting a color from the theme to `themeColor` for resolving ambiguity with the color prop in css
 
 ## [0.12.7] - 2019-08-21